# backend/openapi_server/models/project_snapshot.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server.models.ft_m_entity import FtMEntity
from openapi_server import util

from openapi_server.models.ft_m_entity import FtMEntity  # noqa: E501

class ProjectSnapshot(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, entities=None):  # noqa: E501
        """ProjectSnapshot - a model defined in OpenAPI

        :param entities: The entities of this ProjectSnapshot.  # noqa: E501
        :type entities: List[FtMEntity]
        """
        self.openapi_types = {
            'entities': List[FtMEntity],
        }

        self.attribute_map = {
            'entities': 'entities',
        }

        self._entities = entities

    @classmethod
    def from_dict(cls, dikt) -> 'ProjectSnapshot':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ProjectSnapshot of this ProjectSnapshot.  # noqa: E501
        :rtype: ProjectSnapshot
        """
        return util.deserialize_model(dikt, cls)

    @property
    def entities(self) -> List[FtMEntity]:
        """Gets the entities of this ProjectSnapshot.

        All FtM entities (nodes and interstitial edges) in the graph.  # noqa: E501

        :return: The entities of this ProjectSnapshot.
        :rtype: List[FtMEntity]
        """
        return self._entities

    @entities.setter
    def entities(self, entities: List[FtMEntity]):
        """Sets the entities of this ProjectSnapshot.

        All FtM entities (nodes and interstitial edges) in the graph.  # noqa: E501

        :param entities: The entities of this ProjectSnapshot.
        :type entities: List[FtMEntity]
        """
        if entities is None:
            raise ValueError("Invalid value for `entities`, must not be `None`")  # noqa: E501

        self._entities = entities