# backend/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: |
    api spec for StoryNet, covering the backend API and the internal apis of the pdf_service and ner_service. all public endpoints require JWT authentication unless explicitly marked otherwise.
  title: StoryNet API
  version: 2.0.0
servers:
#- description: backend API
#  url: https://angles-server.ddns.net:5001/api
- description: Internal pdf_service (OCR) — not publicly reachable
  url: http://pdf-service.app.svc.cluster.local:5001
- description: Internal ner_service (NER/LLM) — not publicly reachable
  url: http://ner-service.app.svc.cluster.local:5000
security:
- bearerAuth: []
...
paths:
...
  /projects:
    get:
      operationId: projects_get
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: page_size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
          description: Paginated project list.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
      summary: List all projects belonging to the authenticated user
      tags:
      - Projects
      x-openapi-router-controller: openapi_server.controllers.projects_controller
    post:
      description: |
        Saves the complete current graph state (entities, relationships) as a new project in MongoDB. The snapshot is stored in full so the user sees the exact same view on reload.
      operationId: projects_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_projects_post_request"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
          description: Project created.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Validation error.
      summary: Create and save a new project snapshot
      tags:
      - Projects
      x-openapi-router-controller: openapi_server.controllers.projects_controller
  /projects/{id}:
    delete:
      description: Permanently removes the project from MongoDB. Only the owner may
        delete.
      operationId: projects_id_delete
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Project deleted.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authenticated but not permitted to access this resource.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found or not visible to this user.
      summary: Delete a project
      tags:
      - Projects
      x-openapi-router-controller: openapi_server.controllers.projects_controller
    get:
      description: |
        Returns the full project including the snapshot. The frontend should render the snapshot immediately, then optionally call GET /entities/{id} for each entity to detect changes since the snapshot was taken.
      operationId: projects_id_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
          description: Project loaded.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authenticated but not permitted to access this resource.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found or not visible to this user.
      summary: Load a saved project by ID
      tags:
      - Projects
      x-openapi-router-controller: openapi_server.controllers.projects_controller
    put:
      description: |
        Replaces the stored snapshot entirely with a new one. Used when the user saves changes to an existing project. The previous snapshot is not retained (no versioning).
      operationId: projects_id_put
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_projects__id__put_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
          description: Project updated.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authenticated but not permitted to access this resource.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found or not visible to this user.
      summary: Overwrite an existing project snapshot
      tags:
      - Projects
      x-openapi-router-controller: openapi_server.controllers.projects_controller
...
components:
...
  schemas:
    FtMEntity:
      example:
        schema: Person
        provenance:
        - page_number: 1
          document_id: document_id
        - page_number: 1
          document_id: document_id
        first_seen: 2000-01-23T04:56:07.000+00:00
        public: false
        confidence: 0.87
        caption: caption
        id: id
        properties:
          name:
          - John Doe
          nationality:
          - us
          - au
          birthDate:
          - "1982"
        owner_ids:
        - owner_ids
        - owner_ids
        last_changed: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Stable unique identifier (hash of canonical values).
          title: id
          type: string
        schema:
          description: |
            FtM core schema type. Node-like examples: Person, Organization, Company, LegalEntity, Asset, Vehicle, Vessel, Aircraft, Address, Passport, BankAccount. Edge-like examples: Ownership, Membership, Directorship, Family, Associate, UnknownLink, Sanction, Payment. This is not an exhaustive list; any valid FtM core schema is accepted.
          example: Person
          title: schema
          type: string
        caption:
          description: |
            Human-readable label computed from featured properties (e.g. name). Read-only; computed by the backend.
          title: caption
          type: string
        properties:
          additionalProperties:
            items:
              type: string
            type: array
          description: |
            Multi-valued property bag. ALL values are arrays of strings, even scalar-feeling fields like name or birthDate. Property names are defined by the FtM schema for this entity type. Entity-typed properties (e.g. Ownership.owner) hold the IDs of other entities.
          example:
            name:
            - John Doe
            nationality:
            - us
            - au
            birthDate:
            - "1982"
          title: properties
          type: object
        confidence:
          description: |
            NER confidence score assigned by the ner_service. Stored for all entities regardless of value. Null only for manually created entities.
          example: 0.87
          format: float
          maximum: 1.0
          minimum: 0.0
          title: confidence
          type: number
        public:
          default: false
          description: |
            True if any owning user has released this entity publicly. Derived from the public flags of all documents it was sourced from.
          title: public
          type: boolean
        owner_ids:
          description: User IDs who own at least one source document for this entity.
          items:
            type: string
          title: owner_ids
          type: array
        provenance:
          description: |
            All document+page references that contributed to this entity. An entity sourced from multiple documents will have multiple provenance entries.
          items:
            $ref: "#/components/schemas/ProvenanceEntry"
          title: provenance
          type: array
        first_seen:
          format: date-time
          readOnly: true
          title: first_seen
          type: string
        last_changed:
          format: date-time
          readOnly: true
          title: last_changed
          type: string
      required:
      - id
      - properties
      - schema
      title: FtMEntity
      type: object
    ProvenanceEntry:
      example:
        page_number: 1
        document_id: document_id
      properties:
        document_id:
          title: document_id
          type: string
        page_number:
          description: 1-indexed page number. Null for plain-text documents.
          minimum: 1
          nullable: true
          title: page_number
          type: integer
      required:
      - document_id
      title: ProvenanceEntry
      type: object
...
    Project:
      example:
        updated_at: 2000-01-23T04:56:07.000+00:00
        owner_id: owner_id
        name: name
        description: description
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        snapshot:
          entities:
          - schema: Person
            provenance:
            - page_number: 1
              document_id: document_id
            - page_number: 1
              document_id: document_id
            first_seen: 2000-01-23T04:56:07.000+00:00
            public: false
            confidence: 0.87
            caption: caption
            id: id
            properties:
              name:
              - John Doe
              nationality:
              - us
              - au
              birthDate:
              - "1982"
            owner_ids:
            - owner_ids
            - owner_ids
            last_changed: 2000-01-23T04:56:07.000+00:00
          - schema: Person
            provenance:
            - page_number: 1
              document_id: document_id
            - page_number: 1
              document_id: document_id
            first_seen: 2000-01-23T04:56:07.000+00:00
            public: false
            confidence: 0.87
            caption: caption
            id: id
            properties:
              name:
              - John Doe
              nationality:
              - us
              - au
              birthDate:
              - "1982"
            owner_ids:
            - owner_ids
            - owner_ids
            last_changed: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          title: id
          type: string
        name:
          title: name
          type: string
        description:
          nullable: true
          title: description
          type: string
        owner_id:
          title: owner_id
          type: string
        snapshot:
          $ref: "#/components/schemas/ProjectSnapshot"
        created_at:
          format: date-time
          readOnly: true
          title: created_at
          type: string
        updated_at:
          format: date-time
          readOnly: true
          title: updated_at
          type: string
      required:
      - id
      - name
      - owner_id
      - snapshot
      title: Project
      type: object
    ProjectSnapshot:
      description: |
        Complete graph state saved at project-save time. Entities include both node-type and edge-type FtM entities. The frontend renders this without hitting the live graph database, preserving the exact saved view.
      example:
        entities:
        - schema: Person
          provenance:
          - page_number: 1
            document_id: document_id
          - page_number: 1
            document_id: document_id
          first_seen: 2000-01-23T04:56:07.000+00:00
          public: false
          confidence: 0.87
          caption: caption
          id: id
          properties:
            name:
            - John Doe
            nationality:
            - us
            - au
            birthDate:
            - "1982"
          owner_ids:
          - owner_ids
          - owner_ids
          last_changed: 2000-01-23T04:56:07.000+00:00
        - schema: Person
          provenance:
          - page_number: 1
            document_id: document_id
          - page_number: 1
            document_id: document_id
          first_seen: 2000-01-23T04:56:07.000+00:00
          public: false
          confidence: 0.87
          caption: caption
          id: id
          properties:
            name:
            - John Doe
            nationality:
            - us
            - au
            birthDate:
            - "1982"
          owner_ids:
          - owner_ids
          - owner_ids
          last_changed: 2000-01-23T04:56:07.000+00:00
      properties:
        entities:
          description: All FtM entities (nodes and interstitial edges) in the graph.
          items:
            $ref: "#/components/schemas/FtMEntity"
          title: entities
          type: array
      required:
      - entities
      title: ProjectSnapshot
      type: object
...
    ProjectListResponse:
      example:
        total: 0
        results:
        - updated_at: 2000-01-23T04:56:07.000+00:00
          owner_id: owner_id
          name: name
          description: description
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
          snapshot:
            entities:
            - schema: Person
              provenance:
              - page_number: 1
                document_id: document_id
              - page_number: 1
                document_id: document_id
              first_seen: 2000-01-23T04:56:07.000+00:00
              public: false
              confidence: 0.87
              caption: caption
              id: id
              properties:
                name:
                - John Doe
                nationality:
                - us
                - au
                birthDate:
                - "1982"
              owner_ids:
              - owner_ids
              - owner_ids
              last_changed: 2000-01-23T04:56:07.000+00:00
            - schema: Person
              provenance:
              - page_number: 1
                document_id: document_id
              - page_number: 1
                document_id: document_id
              first_seen: 2000-01-23T04:56:07.000+00:00
              public: false
              confidence: 0.87
              caption: caption
              id: id
              properties:
                name:
                - John Doe
                nationality:
                - us
                - au
                birthDate:
                - "1982"
              owner_ids:
              - owner_ids
              - owner_ids
              last_changed: 2000-01-23T04:56:07.000+00:00
        - updated_at: 2000-01-23T04:56:07.000+00:00
          owner_id: owner_id
          name: name
          description: description
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
          snapshot:
            entities:
            - schema: Person
              provenance:
              - page_number: 1
                document_id: document_id
              - page_number: 1
                document_id: document_id
              first_seen: 2000-01-23T04:56:07.000+00:00
              public: false
              confidence: 0.87
              caption: caption
              id: id
              properties:
                name:
                - John Doe
                nationality:
                - us
                - au
                birthDate:
                - "1982"
              owner_ids:
              - owner_ids
              - owner_ids
              last_changed: 2000-01-23T04:56:07.000+00:00
            - schema: Person
              provenance:
              - page_number: 1
                document_id: document_id
              - page_number: 1
                document_id: document_id
              first_seen: 2000-01-23T04:56:07.000+00:00
              public: false
              confidence: 0.87
              caption: caption
              id: id
              properties:
                name:
                - John Doe
                nationality:
                - us
                - au
                birthDate:
                - "1982"
              owner_ids:
              - owner_ids
              - owner_ids
              last_changed: 2000-01-23T04:56:07.000+00:00
      properties:
        total:
          title: total
          type: integer
        results:
          items:
            $ref: "#/components/schemas/Project"
          title: results
          type: array
      title: ProjectListResponse
      type: object
...
    ErrorResponse:
      example:
        details:
          key: ""
        error: ENTITY_NOT_FOUND
        message: message
      properties:
        error:
          description: Machine-readable error code.
          example: ENTITY_NOT_FOUND
          title: error
          type: string
        message:
          description: Human-readable description.
          title: message
          type: string
        details:
          additionalProperties: true
          nullable: true
          title: details
          type: object
      required:
      - error
      - message
      title: ErrorResponse
      type: object
...
    _projects_post_request:
      properties:
        name:
          title: name
          type: string
        description:
          nullable: true
          title: description
          type: string
        snapshot:
          $ref: "#/components/schemas/ProjectSnapshot"
      required:
      - name
      - snapshot
      title: _projects_post_request
      type: object
    _projects__id__put_request:
      properties:
        name:
          title: name
          type: string
        description:
          nullable: true
          title: description
          type: string
        snapshot:
          $ref: "#/components/schemas/ProjectSnapshot"
      title: _projects__id__put_request
      type: object
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: openapi_server.controllers.security_controller.info_from_bearerAuth